# api-rest-amazonia/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "description": "🔓 **Acceso público — sin token requerido.**\n\nValida email y contraseña y devuelve un token JWT de acceso.\n\n**Uso del token:** incluye el `accessToken` recibido en todas las peticiones protegidas:\n`Authorization: Bearer <accessToken>`\n\n**Seguridad:**\n- El tiempo de respuesta es constante independientemente de si el email existe, para prevenir enumeración de usuarios (OWASP A07).\n- **Rate limiting:** máximo 5 intentos por IP cada 60 segundos. Superar este límite devuelve `429`.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login exitoso. Devuelve el token JWT y su tiempo de expiración.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas, cuenta desactivada o acceso expirado (investigadores)."
          },
          "429": {
            "description": "Demasiadas solicitudes desde esta IP. Espera antes de reintentar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "Auth",
          "Auth — Público"
        ]
      }
    },
...
    "/api/proyectos": {
      "get": {
        "description": "Token opcional. Sin token → devuelve datos resumidos (card). Con token válido → devuelve objeto completo con todas las relaciones.",
        "operationId": "ProyectosController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de la página (empieza en 1). Si no se envía, se toma por defecto el valor 1.",
            "schema": {
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de resultados por página. Si no se envía, se toma por defecto el valor 10.",
            "schema": {
              "maximum": 100,
              "default": 10,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "area",
            "required": false,
            "in": "query",
            "description": "Filtrar por área: 1 = Conservación, 2 = Desarrollo Comunitario",
            "schema": {
              "example": 1,
              "type": "number",
              "enum": [
                1,
                2
              ]
            }
          },
          {
            "name": "departamento",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de departamento donde opera el proyecto",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de tipo de proyecto",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "anio",
            "required": false,
            "in": "query",
            "description": "Filtrar por año de inicio del proyecto",
            "schema": {
              "example": 2023,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nombre de proyecto (búsqueda parcial, insensible a mayúsculas)",
            "schema": {
              "example": "Reforestación",
              "type": "string"
            }
          },
          {
            "name": "municipio",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de municipio de trabajo del proyecto",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "anio_desde",
            "required": false,
            "in": "query",
            "description": "Filtrar proyectos que iniciaron desde este año (inclusive)",
            "schema": {
              "example": 2020,
              "type": "number"
            }
          },
          {
            "name": "anio_hasta",
            "required": false,
            "in": "query",
            "description": "Filtrar proyectos que terminaron hasta este año (inclusive). Incluye proyectos activos (sin anio_fin)",
            "schema": {
              "example": 2024,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Ordenar resultados. Formato: 'campo:asc' o 'campo:desc'. Campos válidos: nombre, anioInicio",
            "schema": {
              "example": "anioInicio:desc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado. Shape varía según autenticación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar proyectos con filtros y paginación",
        "tags": [
          "Proyectos",
          "Proyectos — Público"
        ]
      }
    },
...
    "/api/proyectos/{id}": {
      "get": {
        "operationId": "ProyectosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proyecto con todas sus relaciones, galería e imagen principal"
          },
          "400": {
            "description": "Respuesta en caso de parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationExceptionDto"
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido — {\"error\":\"AUTHENTICATION_REQUIRED\",\"message\":\"...\",\"login_url\":\"/api/auth/login\"}"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Detalle completo de un proyecto. Requiere token JWT.",
        "tags": [
          "Proyectos",
          "Proyectos — Autenticado"
        ]
      }
    },
    "/api/proyectos/{id}/imagen-principal": {
      "post": {
        "operationId": "ProyectosController_uploadImagenPrincipal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ imagenPrincipalUrl: string }"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Subir o reemplazar imagen principal del proyecto (máx 5 MB, JPEG/PNG/WebP → WebP 1200×800)",
        "tags": [
          "Proyectos",
          "Proyectos — Autenticado"
        ]
      },
      "delete": {
        "operationId": "ProyectosController_deleteImagenPrincipal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Imagen eliminada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Eliminar imagen principal del proyecto",
        "tags": [
          "Proyectos",
          "Proyectos — Autenticado"
        ]
      }
    },
    "/api/proyectos/{id}/galeria": {
      "post": {
        "operationId": "ProyectosController_uploadGaleria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "descripcion": {
                    "type": "string",
                    "description": "Descripción opcional de la imagen"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen de galería creada con id, url y orden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Añadir imagen a la galería del proyecto (máx 5 MB, JPEG/PNG/WebP → WebP 1200×800)",
        "tags": [
          "Proyectos",
          "Proyectos — Autenticado"
        ]
      }
    },
    "/api/proyectos/{id}/galeria/{imagenId}": {
      "delete": {
        "operationId": "ProyectosController_deleteGaleriaImagen",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "imagenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Imagen eliminada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Eliminar imagen de la galería",
        "tags": [
          "Proyectos",
          "Proyectos — Autenticado"
        ]
      }
    },
    "/api/proyectos/{id}/galeria/orden": {
      "put": {
        "operationId": "ProyectosController_reordenarGaleria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orden actualizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "🔒 Reordenar imágenes de la galería. Body: [{id: string, orden: number}]",
        "tags": [
          "Proyectos",
          "Proyectos — Autenticado"
        ]
      }
    },
...
    "/api/dashboard/proyectos/{id}": {
      "get": {
        "description": "Devuelve todos los datos de un proyecto por su ID. La estructura del objeto varía según el área:\n\n- **Conservación** (id_area=1): `especies_animales` y `practicas_agricolas` contienen arrays con los elementos vinculados. `areas_desarrollo` es siempre `null`.\n- **Desarrollo** (id_area=2): `areas_desarrollo` contiene el array de áreas de desarrollo comunitario. `especies_animales` y `practicas_agricolas` son siempre `null`.\n\nPara construir la vista de detalle del frontend, verificar `id_area` antes de renderizar las secciones específicas de cada área.\n\n_Caché: 120 segundos._",
        "operationId": "DashboardController_getProyectoById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Objeto proyecto completo. El campo id_area determina qué sección de datos específicos está poblada.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id_proyecto": 1,
                    "nombre": "Conservación de jaguar en el Beni",
                    "descripcion": "Proyecto de monitoreo y conservación de jaguar (Panthera onca) en la llanura beniana.",
                    "anio_inicio": 2021,
                    "anio_fin": null,
                    "activo": true,
                    "id_area": 1,
                    "area": "Conservacion",
                    "id_tipo": 6,
                    "tipo_proyecto": "Conservación de especies",
                    "total_empresas_participantes": 2,
                    "total_organizaciones_participantes": 1,
                    "departamentos": [
                      {
                        "id": 2,
                        "nombre": "Beni",
                        "amazonico": true
                      }
                    ],
                    "municipios": [
                      {
                        "id": 49,
                        "nombre": "Trinidad"
                      }
                    ],
                    "comunidades_indigenas": [
                      {
                        "id": 22,
                        "nombre": "Moxeño"
                      }
                    ],
                    "tipos_ayuda": [
                      {
                        "id": 2,
                        "nombre": "Financiero"
                      }
                    ],
                    "actores_locales": [
                      {
                        "id": 1,
                        "nombre": "Gobiernos municipales"
                      }
                    ],
                    "especies_animales": [
                      {
                        "id": 1,
                        "nombre": "Jaguar"
                      },
                      {
                        "id": 2,
                        "nombre": "Bufeo"
                      }
                    ],
                    "practicas_agricolas": [
                      {
                        "id": 2,
                        "nombre": "Castaña"
                      }
                    ],
                    "areas_desarrollo": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido, expirado o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Respuesta en caso de no encontrar un recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detalle completo de un proyecto amazónico",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kaa Iya — Backend API",
    "description": "API REST de la plataforma **Kaa Iya** (\"Espíritu del Bosque\"), para gestionar y visibilizar iniciativas sostenibles en la Amazonía boliviana.\n\n**Base URL:** `/api`\n\n**Autenticación:** Bearer JWT. Obtén tu token en `POST /auth/login` e inclúyelo en todas las peticiones protegidas: `Authorization: Bearer <token>`",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Kaa Iya — UCB San Pablo",
      "url": "https://www.ucb.edu.bo",
      "email": "catedra.amazonia@ucb.edu.bo"
    },
    "license": {
      "name": "UNLICENSED",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development"
    },
    {
      "url": "https://kaaiya-backend.onrender.com",
      "description": "Producción (Render/Railway — ajustar al dominio real)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT obtenido de POST /auth/login"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico del usuario",
            "example": "admin@kaaiya.bo"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "MiPassword1!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Token JWT de acceso",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "tipo": {
            "type": "string",
            "description": "Tipo de token",
            "example": "Bearer"
          },
          "expiresIn": {
            "type": "string",
            "description": "Tiempo de expiración del token",
            "example": "24h"
          }
        },
        "required": [
          "accessToken",
          "tipo",
          "expiresIn"
        ]
      },
      "ValidationExceptionDto": {
        "type": "object",
        "properties": {
          "message": {
            "description": "Lista de errores a corregir",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "description": "Nombre del error"
          },
          "statusCode": {
            "type": "number",
            "description": "Codigo de error"
          }
        },
        "required": [
          "message",
          "error",
          "statusCode"
        ]
      },
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje de respuesta"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "PaginationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array de elementos de la página actual",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "page": {
            "type": "number",
            "description": "Número de la página actual",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Cantidad de elementos por página",
            "example": 10
          },
          "pages": {
            "type": "number",
            "description": "Cantidad total de páginas disponibles",
            "example": 5
          },
          "total": {
            "type": "number",
            "description": "Cantidad total de registros encontrados",
            "example": 42
          },
          "has_next": {
            "type": "boolean",
            "description": "Indica si existe una página siguiente",
            "example": true
          },
          "has_prev": {
            "type": "boolean",
            "description": "Indica si existe una página anterior",
            "example": false
          }
        },
        "required": [
          "data",
          "page",
          "limit",
          "pages",
          "total",
          "has_next",
          "has_prev"
        ]
      },
...
    }
  }
}